Tests/LibWeb/Text/expected/dialog-light-dismiss.txt

closerequest="Any" should respond to light dismissal
other dialogs should not respond to light dismissal
PASS
